test(env,devices): add comprehensive integration and unit tests
- Add extensive integration tests for SpinTorqueEnv, SpinTorqueArrayEnv, and SkyrmionRacetrackEnv covering creation, reset, step, action and observation spaces, reward bounds, episode termination, deterministic behavior, and error handling.
- Add unit tests for STTMRAMDevice, SOTMRAMDevice, VCMAMRAMDevice, SkyrmionDevice, and DeviceFactory covering initialization, parameter validation, effective field computation, resistance calculation, switching thresholds, power consumption, and other device-specific behaviors.
- Improve test coverage and robustness for device models and environment implementations.
- Refactor and expand existing tests to ensure compliance with expected behaviors and error handling.

+++ spin_torque_gym/devices/stt_mram.py
@@ -60,7 +60,7 @@ class STTMRAMDevice(BaseSpintronicDevice):
         m = self.validate_magnetization(magnetization)
         
         # Applied field contribution
-        h_eff = applied_field.copy()
+        h_eff = applied_field.astype(float).copy()
         
         # Uniaxial anisotropy field
         k_u = self.get_parameter('uniaxial_anisotropy', 1e6)
